ref: Move test file into JsPsych folder and delete old images
